+++ src/carve/BuiltinCarvers.cpp
@@ -4,6 +4,7 @@
 #include <memory>
 
 #include "formats/JpegCarver.hpp"
+#include "formats/Mp4Carver.hpp"
 #include "formats/PngCarver.hpp"
 #include "revenant/carve/CarverRegistry.hpp"
 
@@ -12,6 +13,7 @@ namespace revenant::carve {
 void registerBuiltinCarvers(CarverRegistry& registry) {
 	registry.registerCarver(std::make_unique<JpegCarver>());
 	registry.registerCarver(std::make_unique<PngCarver>());
+	registry.registerCarver(std::make_unique<Mp4Carver>());
 }
 
 } // namespace revenant::carve